ContractCoverageCollector
